Create README.md
This Project is an Auto-Drum-Player, consisting of a board that has some cells and these cells represent a musical note.
The user has the opportunity to choose what component of the Drum to play and do so by clicking on one of the cells or in as many as they wish to create their rhythm which will be coloured. 
Once chosen the user then can click on the play which will start a reader ( The reader - Is a shadow that will be applied to the column of the cells and at the moment that the reader pass through the columns whatever cells which are clicked then they are supposed to reproduce a sound prevenient from each part of the drum and this way create your own "music".
for those users who never had the experience of playing there is a set of preset rhythms that will help to get the understanding.
The project is based on a School of Music app.